Toggle page creation based on page and card count
After adding a card to a page, we pre-create the next page if the
current one is full. To avoid an empty page, we would first check
if the card we added was the last one left. But we did this based
on file path, so we risked prematurely exiting if there were still
more of that same card to add.

Now, we calculate how many pages are needed to fit all the cards,
so we can infer that we're on the last card when the final page is
full.